Factor out sentry CLI call method to helper
In preperation for adding more actions, factor out common code.
